requirements.txt
Flask>=0.10.1
kazoo
GitPython>=0.3.2.RC1
pytz
IPython
pygments